requirements.txt
asgiref==3.6.0
cfgv==3.3.1
distlib==0.3.6
Django==4.2
djangorestframework==3.14.0
filelock==3.12.0
flake8==6.0.0
identify==2.5.22
mccabe==0.7.0
nodeenv==1.7.0
platformdirs==3.2.0
pre-commit==3.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
virtualenv==20.22.0
psycopg2==2.9.6
drf-spectacular>=0.26.2
Pillow==9.5.0
django-simple-history==3.3.0